Updated files  to allow multuuser functionality, new.ejs+new.css

+++ index.js
@@ -10,7 +10,7 @@ const port = 3000;
 const db = new pg.Client({
   user: 'postgres',
   host: 'localhost',
-  password: '', //enter your postgres pw here
+  password: 'H72PiSwS!', //enter your postgres pw here
   database: 'world',
   port: 5432,
 });
@@ -21,13 +21,32 @@ db.connect();
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(express.static("public"));
 
+async function user_list() {
+  const search_result = await db.query(
+    "SELECT * FROM users"
+  );
+  let user_list = search_result.rows;
+  return user_list
+}
+
+let tempList = await user_list();
+let currentUserId = tempList[0].id;
+
+async function current_user() {
+  const search_result = await db.query(
+    "SELECT * FROM users WHERE users.id = $1", [currentUserId]
+  );
+  let user_list = search_result.rows;
+  return user_list[0];
+}
+
 //function to get the previously entered countries from the db
 async function country_list() {
   const search_result = await db.query(
-    "SELECT country_code FROM visited_countries;"
+    "SELECT country_code FROM visited_countries AS v JOIN users AS u ON u.id = v.user_id WHERE u.id=$1",
+    [currentUserId]
   );
   let visited_countries = search_result.rows;
-
   let countries_count = 0;
   let countries_array = [];
   visited_countries.forEach((element) => {
@@ -40,17 +59,28 @@ async function country_list() {
 
 //home page
 app.get("/", async (req, res) => {
+  let currentUser = await current_user();
   let countries = await country_list();
+  let users = await user_list();
 
+  if (currentUserId) {
   res.render("index.ejs", {
     total: countries[1],
-    countries: countries[0]
+    countries: countries[0],
+    users: users,
+    color: currentUser.color,
   });
+  } else {
+    res.render('new.ejs');
+  }
 });
 
 //adding or removing a country from the list
 app.post("/change", async (req,res) => {
   let entered_data = req.body.country.toLowerCase().trim();
+  let currentUser = await current_user();
+  let users = await user_list();
+
   //checking if the entered country exist
   try {
     let first_query = await db.query(
@@ -63,7 +93,7 @@ app.post("/change", async (req,res) => {
       //adding the entered country to the list
       try {
         await db.query(
-          "INSERT INTO visited_countries (country_code) Values($1)", [selected_code]
+          "INSERT INTO visited_countries (country_code, user_id) Values($1, $2)", [selected_code, currentUserId]    //CHANGE THIS TO ADD USER_id
         );
         res.redirect('/');
       } 
@@ -76,12 +106,14 @@ app.post("/change", async (req,res) => {
         res.render("index.ejs", {
           total: countries[1],
           countries: countries[0],
+          users: users,
+          color: currentUser.color,
           error: "The country has already been added. Try again"
         });
       };
       //removing entered country from the list
     } else {
-      await db.query("DELETE FROM visited_countries WHERE country_code=$1", [selected_code]);
+      await db.query("DELETE FROM visited_countries WHERE country_code=$1 AND user_id=$2", [selected_code, currentUserId]);
       res.redirect('/');
     }
 
@@ -94,6 +126,8 @@ app.post("/change", async (req,res) => {
     res.render("index.ejs", {
       total: countries[1],
       countries: countries[0],
+      users:users,
+      color: currentUser.color,
       error: "The country you've entered does not exist. Please, try again"
     });
   };
@@ -102,10 +136,50 @@ app.post("/change", async (req,res) => {
 
 //delete all countries from the current list
 app.get("/delete", async (req, res) => {
-  await db.query("DELETE FROM visited_countries");
+  await db.query("DELETE FROM visited_countries WHERE user_id=$1", [currentUserId]);
   res.redirect('/');
 });
 
+app.post("/user", async (req, res) => {
+  if (req.body.add !== 'new') {
+    currentUserId = req.body.user;
+    res.redirect('/');
+  }
+  else {
+    res.render('new.ejs')
+  }
+});
+
+app.post("/new", async (req, res) => {
+  let newName = req.body.name;
+  let newColor = req.body.color;
+  await db.query(
+    "INSERT INTO users (name, color) VALUES ($1, $2)", [newName, newColor]
+  );
+  tempList = await user_list();
+  currentUserId = tempList[0].id;
+  res.redirect('/');
+});
+
+
+app.post("/deleteUser", async (req, res) => {
+  tempList = await user_list();
+
+  let deleteValue = req.body.deleteUser;
+  await db.query(
+    "DELETE FROM visited_countries WHERE user_id = $1", [deleteValue]
+  );
+  await db.query(
+    "DELETE FROM users WHERE id = $1", [deleteValue]
+  );
+  if (tempList.length === 1) {  
+    res.render("new.ejs")
+  } else {
+  tempList = await user_list();
+  currentUserId = tempList[0].id;
+  res.redirect('/');
+  }
+});
 
 app.listen(port, () => {
   console.log(`Server running on http://localhost:${port}`);
